+++ R/gitstats_functions.R
@@ -260,9 +260,13 @@ get_commits <- function(gitstats_object,
 #' @description Prepare statistics from the pulled commits data.
 #' @details To make function work, you need first to get commits data with
 #'   `GitStats`. See examples section.
-#' @param gitstats_object A GitStats class object.
-#' @param time_interval A character, specifying time interval to show
+#' @param gitstats_object A `GitStats` object.
+#' @param time_aggregation A character, specifying time aggregation of
 #'   statistics.
+#' @param group_var Other grouping variable to be passed to `dplyr::group_by()`
+#'   function apart from `stats_date` and `githost`. Could be: `author`,
+#'   `author_login`, `author_name` or `organization`. Should be passed without
+#'   quotation marks.
 #' @return A table of `commits_stats` class.
 #' @examples
 #' \dontrun{
@@ -272,13 +276,20 @@ get_commits <- function(gitstats_object,
 #'      repos = c("r-world-devs/GitStats", "openpharma/visR")
 #'    )
 #'  get_commits(my_gitstats, since = "2022-01-01")
-#'  get_commits_stats(my_gitstats, time_interval = "week")
+#'  get_commits_stats(
+#'    gitstats_object = my_gitstats,
+#'    group_var = author,
+#'    time_interval = "year"
+#'  )
 #' }
 #' @export
 get_commits_stats <- function(gitstats_object,
-                              time_interval = c("month", "day", "week")) {
+                              time_aggregation = c("year", "month", "week", "day"),
+                              group_var) {
+  time_aggregation <- match.arg(time_aggregation)
   gitstats_object$get_commits_stats(
-    time_interval = time_interval
+    time_aggregation = time_aggregation,
+    group_var = rlang::enquo(group_var)
   )
 }