+++ tests/testthat/test-get_commits-GitStats.R
@@ -74,10 +74,34 @@ test_that("get_commits() returns error when since is not defined", {
   )
 })
 
+test_that("prepare_commits_stats prepares commits statistics", {
+  commits_stats <- test_gitstats_priv$prepare_commits_stats(
+    commits = test_mocker$use("commits_table"),
+    time_aggregation = "week",
+    group_var = author
+  )
+  expect_equal(
+    colnames(commits_stats),
+    c("stats_date", "githost", "author", "stats")
+  )
+})
+
 test_gitstats <- create_test_gitstats(
   hosts = 2,
   inject_commits = "commits_table"
 )
+
+test_that("get_commits_stats method works", {
+  commits_stats <- test_gitstats$get_commits_stats(
+    time_aggregation = "month",
+    group_var = "organization"
+  )
+  expect_s3_class(commits_stats, "commits_stats")
+  expect_true(
+    all(c("gitlab", "github") %in% commits_stats$githost)
+  )
+  test_mocker$cache(commits_stats)
+})
 
 test_that("get_commits_stats returns error when no commits", {
   test_gitstats <- create_test_gitstats()
@@ -87,23 +111,26 @@ test_that("get_commits_stats returns error when no commits", {
 })
 
 test_that("get_commits_stats prepares table with statistics on commits", {
-  commits_stats <- get_commits_stats(test_gitstats)
-  expect_s3_class(commits_stats, "commits_stats")
-  expect_equal(
-    colnames(commits_stats),
-    c("stats_date", "platform", "organization", "commits_n")
-  )
-  expect_true(
-    "github" %in% commits_stats$platform
-  )
-  test_mocker$cache(commits_stats)
-
   commits_stats_daily <- get_commits_stats(
     gitstats_obj = test_gitstats,
-    time_interval = "day")
+    time_aggregation = "day",
+    group_var = organization
+  )
   expect_s3_class(commits_stats_daily, "commits_stats")
   expect_equal(
     colnames(commits_stats_daily),
-    c("stats_date", "platform", "organization", "commits_n")
+    c("stats_date", "githost", "organization", "stats")
+  )
+
+  commits_stats_yearly <- get_commits_stats(
+    gitstats_obj = test_gitstats,
+    time_aggregation = "year"
+  )
+  expect_equal(commits_stats_yearly$stats_date,
+               as.POSIXct(c(rep("2023-01-01", 2), "2024-01-01"), tz = 'UTC'))
+  expect_s3_class(commits_stats_yearly, "commits_stats")
+  expect_equal(
+    colnames(commits_stats_yearly),
+    c("stats_date", "githost", "stats")
   )
 })